+++ Src/C/run-time/eif_debug.h
@@ -196,12 +196,12 @@ extern void drecord_bc(BODY_INDEX body_idx, BODY_INDEX body_id, unsigned char *a
 extern void ewhere(struct where *where);
 
 /* frozen stack (used to record local variables address) */ 
-extern EIF_TYPED_ADDRESS	*c_stack_allocate(register int size);
+extern EIF_TYPED_ADDRESS	*c_stack_allocate(register size_t size);
 extern EIF_TYPED_ADDRESS	*c_opush(register EIF_TYPED_ADDRESS *val);
 extern EIF_TYPED_ADDRESS	*c_opop(void);
 extern EIF_TYPED_ADDRESS	*c_otop(void);
 extern EIF_TYPED_ADDRESS 	*c_oitem(uint32 n);
-extern int			c_stack_extend(register int size);
+extern int			c_stack_extend(register size_t size);
 extern void 		c_npop(register int nb_items);
 extern void			c_wipe_out(register struct c_stochunk *chunk);
 extern void 		c_stack_truncate(void);
