+++ Src/C/run-time/eif_local.h
@@ -43,7 +43,7 @@ extern "C" {
 
 RT_LNK void epop(struct stack *stk, rt_uint_ptr nb_items);			/* Pops values off the local stack */
 #ifdef ISE_GC
-RT_LNK char **eget(register int num);		/* Get another chunk for local variables */
+RT_LNK char **eget(register size_t num);		/* Get another chunk for local variables */
 RT_LNK void eback(register char **top);		/* Get back to the previous stack chunk */
 #endif
 RT_LNK void initstk(void);		/* Initialize local stacks */
